Add abstract ProtocolClient/Refactor

+++ src/Toolbox/Protocol/MessageTokens.cs
@@ -1,11 +1,12 @@
 ﻿namespace Toolbox.Protocol
 {
+    // TODO: Convert to ENUM
     public class MessageTokens
     {
-        public const int STX = 2;
-        public const int ETX = 3;
-        public const int ACK = 6;
-        public const int NAK = 21;
-        public const int ESC = 27;
+        public const byte STX = 2;
+        public const byte ETX = 3;
+        public const byte ACK = 6;
+        public const byte NAK = 21;
+        public const byte ESC = 27;
     }
 }
